refactor(sandbox): extract content-blocks and slim DocPreview
Break DocPreview into composable pieces:

content-blocks/ — renders ContentBlock types from docs-types.ts:
  ProseBlock, CodeBlock, TableBlock, ListBlock, ContentBlockRenderer

SectionRenderer — renders a ReferenceSection with merged do/dont lists

DocPreview — thin shell (181 lines, down from 1129):
  title/version, overview/usage/practice sections via SectionRenderer,
  token tables via token-tables/ components routed by doc.tokenCategory

All preview renderers, token table logic, and helpers removed from
DocPreview — each concern lives in its own module.
